# BovineBoss-API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "BovineBoss-API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Trabajador/registerTrabajador": {
      "post": {
        "tags": [
          "Trabajador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateEmployeeDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Trabajador/ActualizarTrabajador": {
      "put": {
        "tags": [
          "Trabajador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ModifyTrabajadorDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Trabajador/AgregarTrabajadorFinca": {
      "post": {
        "tags": [
          "Trabajador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateNewEstateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Trabajador/ActualizarTrabajadorAdmin": {
      "put": {
        "tags": [
          "Trabajador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ModifyTrabajadorAdminDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Trabajador/GetWorkers": {
      "get": {
        "tags": [
          "Trabajador"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "estateId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CreateEmployeeDto": {
      "type": "object",
      "properties": {
        "nombrePersona": {
          "type": "string"
        },
        "apellidoPersona": {
          "type": "string"
        },
        "cedula": {
          "type": "string"
        },
        "telefonoPersona": {
          "type": "string"
        },
        "salario": {
          "format": "int32",
          "type": "integer"
        },
        "usuario": {
          "type": "string"
        },
        "contrasenia": {
          "type": "string"
        },
        "idFinca": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "CreateNewEstateDto": {
      "type": "object",
      "properties": {
        "idEmployee": {
          "format": "int32",
          "type": "integer"
        },
        "listIdState": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        }
      }
    },
...
    "ModifyTrabajadorAdminDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombrePersona": {
          "type": "string"
        },
        "apellidoPersona": {
          "type": "string"
        },
        "cedula": {
          "type": "string"
        },
        "telefonoPersona": {
          "type": "string"
        },
        "usuario": {
          "type": "string"
        },
        "contrasenia": {
          "type": "string"
        },
        "salario": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "ModifyTrabajadorDto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombrePersona": {
          "type": "string"
        },
        "apellidoPersona": {
          "type": "string"
        },
        "cedula": {
          "type": "string"
        },
        "telefonoPersona": {
          "type": "string"
        },
        "usuario": {
          "type": "string"
        },
        "salario": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}